+++ examples/usage_reports_to_adw/setup/setup_packages.sh
@@ -7,7 +7,7 @@
 # Written by Adi Zohar, October 2020
 # Git Location = https://github.com/oracle/oci-python-sdk/tree/master/examples/usage_reports_to_adw
 #
-# Version 2020-10-22
+# Version 2020-12-17
 #
 #########################################################################################################################
 
@@ -27,7 +27,7 @@ echo "########################################################################" 
 echo "# 1. Install Python3, Git and Python OCI Packages, Can take a moment." | tee -a $LOG
 echo "########################################################################" | tee -a $LOG
 sudo yum install -y python3 git | tee -a $LOG
-sudo pip3 install oci oci-cli cx_Oracle requests | tee -a $LOG
+sudo python3 -m pip install oci oci-cli cx_Oracle requests | tee -a $LOG
 echo "Completed." | tee -a $LOG
 
 ###########################################
@@ -35,11 +35,21 @@ echo "Completed." | tee -a $LOG
 ###########################################
 echo "" | tee -a $LOG
 echo "########################################################################" | tee -a $LOG
-echo "# 2. Install Oracle Instant Client." | tee -a $LOG
+echo "# 2. Install Oracle Instant Client 19.9." | tee -a $LOG
 echo "########################################################################" | tee -a $LOG
-sudo yum install -y oracle-release-el7 | tee -a $LOG
-sudo yum install -y oracle-instantclient19.8-basic.x86_64 oracle-instantclient19.8-sqlplus.x86_64 | tee -a $LOG
-echo "Completed." | tee -a $LOG
+sudo rpm -i https://yum.oracle.com/repo/OracleLinux/OL7/oracle/instantclient/x86_64/getPackage/oracle-instantclient19.9-basic-19.9.0.0.0-1.x86_64.rpm | tee -a $LOG
+sudo rpm -i https://yum.oracle.com/repo/OracleLinux/OL7/oracle/instantclient/x86_64/getPackage/oracle-instantclient19.9-sqlplus-19.9.0.0.0-1.x86_64.rpm | tee -a $LOG
+sudo ln -s /usr/lib/oracle/19.9 /usr/lib/oracle/current | tee -a $LOG
+
+# Check if installed
+echo "Check Installation..." | tee -a $LOG
+rpm -qa |grep oracle-instantclient
+if [ $? -eq 0 ]; then
+   echo "   Completed." | tee -a $LOG
+else
+   echo "   Error installing oracle instant client, need to perform it manually, Abort, log=$LOG" | tee -a $LOG
+   exit 1
+fi
 
 ###########################################
 # Setup .bashrc profile
@@ -48,7 +58,7 @@ echo "" | tee -a $LOG
 echo "########################################################################" | tee -a $LOG
 echo "# 3. Setup .bashrc env variables." | tee -a $LOG
 echo "########################################################################" | tee -a $LOG
-echo "export CLIENT_HOME=/usr/lib/oracle/19.8/client64" >>$HOME/.bashrc
+echo "export CLIENT_HOME=/usr/lib/oracle/current/client64" >>$HOME/.bashrc
 echo "export LD_LIBRARY_PATH=$CLIENT_HOME/lib" >>$HOME/.bashrc
 echo "export PATH=$PATH:$CLIENT_HOME/bin" >>$HOME/.bashrc
 echo "export TNS_ADMIN=$HOME/ADWCUSG" >>$HOME/.bashrc