+++ src/oci/bds/models/node.py
@@ -1,5 +1,5 @@
 # coding: utf-8
-# Copyright (c) 2016, 2020, Oracle and/or its affiliates.  All rights reserved.
+# Copyright (c) 2016, 2021, Oracle and/or its affiliates.  All rights reserved.
 # This software is dual-licensed to you under the Universal Permissive License (UPL) 1.0 as shown at https://oss.oracle.com/licenses/upl or Apache License 2.0 as shown at http://www.apache.org/licenses/LICENSE-2.0. You may choose either license.
 
 
@@ -41,14 +41,6 @@ class Node(object):
     #: This constant has a value of "FAILED"
     LIFECYCLE_STATE_FAILED = "FAILED"
 
-    #: A constant which can be used with the lifecycle_state property of a Node.
-    #: This constant has a value of "STOPPING"
-    LIFECYCLE_STATE_STOPPING = "STOPPING"
-
-    #: A constant which can be used with the lifecycle_state property of a Node.
-    #: This constant has a value of "STARTING"
-    LIFECYCLE_STATE_STARTING = "STARTING"
-
     #: A constant which can be used with the node_type property of a Node.
     #: This constant has a value of "MASTER"
     NODE_TYPE_MASTER = "MASTER"
@@ -88,7 +80,7 @@ class Node(object):
 
         :param lifecycle_state:
             The value to assign to the lifecycle_state property of this Node.
-            Allowed values for this property are: "CREATING", "ACTIVE", "INACTIVE", "UPDATING", "DELETING", "DELETED", "FAILED", "STOPPING", "STARTING", 'UNKNOWN_ENUM_VALUE'.
+            Allowed values for this property are: "CREATING", "ACTIVE", "INACTIVE", "UPDATING", "DELETING", "DELETED", "FAILED", 'UNKNOWN_ENUM_VALUE'.
             Any unrecognized values returned by a service will be mapped to 'UNKNOWN_ENUM_VALUE'.
         :type lifecycle_state: str
 
@@ -249,7 +241,7 @@ class Node(object):
         **[Required]** Gets the lifecycle_state of this Node.
         The state of the node
 
-        Allowed values for this property are: "CREATING", "ACTIVE", "INACTIVE", "UPDATING", "DELETING", "DELETED", "FAILED", "STOPPING", "STARTING", 'UNKNOWN_ENUM_VALUE'.
+        Allowed values for this property are: "CREATING", "ACTIVE", "INACTIVE", "UPDATING", "DELETING", "DELETED", "FAILED", 'UNKNOWN_ENUM_VALUE'.
         Any unrecognized values returned by a service will be mapped to 'UNKNOWN_ENUM_VALUE'.
 
 
@@ -268,7 +260,7 @@ class Node(object):
         :param lifecycle_state: The lifecycle_state of this Node.
         :type: str
         """
-        allowed_values = ["CREATING", "ACTIVE", "INACTIVE", "UPDATING", "DELETING", "DELETED", "FAILED", "STOPPING", "STARTING"]
+        allowed_values = ["CREATING", "ACTIVE", "INACTIVE", "UPDATING", "DELETING", "DELETED", "FAILED"]
         if not value_allowed_none_or_none_sentinel(lifecycle_state, allowed_values):
             lifecycle_state = 'UNKNOWN_ENUM_VALUE'
         self._lifecycle_state = lifecycle_state